refactor: effects module
