+++ fury/effects/__init__.py
@@ -1,0 +1,2 @@
+from fury.effects.effect_manager import EffectManager
+from fury.effects.effects import *
